requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlmodel==0.0.14
sqlalchemy==2.0.23

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# HTTP
httpx==0.25.2

# Settings  
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1